files/requirements.txt
# COVID-19 Detection Django Project - Requirements
# Python 3.8+ required

# Django Core
Django==4.2.7
django-crispy-forms==2.1
crispy-bootstrap5==0.7

# Database
psycopg2-binary==2.9.9  # PostgreSQL (optional, can use SQLite)

# Deep Learning & AI
torch==2.1.0
torchvision==0.16.0
timm==0.9.12  # PyTorch Image Models (for CrossViT, ResNet, etc.)
pytorch-grad-cam==1.4.8  # For Grad-CAM explainability

# Image Processing
Pillow==10.1.0
opencv-python==4.8.1.78
albumentations==1.3.1  # For CLAHE and augmentations
scikit-image==0.22.0

# Scientific Computing
numpy==1.24.3
scipy==1.11.4
pandas==2.1.3

# Visualization
matplotlib==3.8.2
seaborn==0.13.0

# Utilities
python-decouple==3.8  # For environment variables
tqdm==4.66.1

# Production (Optional)
gunicorn==21.2.0
whitenoise==6.6.0  # Static files serving

# Development (Optional)
django-debug-toolbar==4.2.0
black==23.12.0
flake8==6.1.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
